+++ CHANGELOG.md
@@ -9,5 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [0.1.0] - 2019-04-22
 ### Added
+- ModalContext, Provider and consumer.
+
+### Added
 
 - **Component** Create the VTEX Store Component Modal